waterBack/requirements.txt
# Web Framework
fastapi==0.116.0
uvicorn==0.30.1
python-multipart==0.0.20

# LangChain
langchain==0.3.26
langchain-openai==0.3.27
langchain-community==0.3.27

# PDF Processing
pypdf==5.7.0
pdfplumber==0.11.5
pdf2image==1.17.0
python-magic-bin==0.4.14

# Pinecone (for potential vector storage)
pinecone-plugin-inference==1.1.0

# PDF Generation
reportlab==4.2.0
weasyprint==62.3
jinja2==3.1.6

# Data Processing
pandas==2.3.1
numpy==2.3.1
python-dateutil==2.9.0.post0

# HTTP & API
requests==2.32.4
httpx==0.28.1
aiofiles==24.1.0

# Utilities
python-dotenv==1.1.1
pydantic==2.8.2
typing-extensions==4.13.2
markdown==3.7

# Development
pytest==8.3.2
pytest-asyncio==0.23.7
black==24.4.2 